Javascripting
Did more of the challenges. Ending the night on the array filtering.
That confused me a little as the example does't explain much. After
looking on MDN i realized it was explicit and i was able to move forward
but the code was a little hard to follow at first the way things are
passed around. I am gonna need to try that one a few more times for it
to sink in.